plot: separate test examples from regular tests according to https://blog.golang.org/examples issue: #545
