+++ plotter/example_functions_test.go
@@ -1,0 +1,55 @@
+// Copyright ©2015 The Gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package plotter_test
+
+import (
+	"gonum.org/v1/plot"
+	"gonum.org/v1/plot/plotter"
+	"gonum.org/v1/plot/vg"
+	"image/color"
+	"log"
+	"math"
+)
+
+// ExampleFunction draws some functions.
+func ExampleFunction() {
+	quad := plotter.NewFunction(func(x float64) float64 { return x * x })
+	quad.Color = color.RGBA{B: 255, A: 255}
+
+	exp := plotter.NewFunction(func(x float64) float64 { return math.Pow(2, x) })
+	exp.Dashes = []vg.Length{vg.Points(2), vg.Points(2)}
+	exp.Width = vg.Points(2)
+	exp.Color = color.RGBA{G: 255, A: 255}
+
+	sin := plotter.NewFunction(func(x float64) float64 { return 10*math.Sin(x) + 50 })
+	sin.Dashes = []vg.Length{vg.Points(4), vg.Points(5)}
+	sin.Width = vg.Points(4)
+	sin.Color = color.RGBA{R: 255, A: 255}
+
+	p, err := plot.New()
+	if err != nil {
+		log.Panic(err)
+	}
+
+	p.Title.Text = "Functions"
+	p.X.Label.Text = "X"
+	p.Y.Label.Text = "Y"
+
+	p.Add(quad, exp, sin)
+	p.Legend.Add("x^2", quad)
+	p.Legend.Add("2^x", exp)
+	p.Legend.Add("10*sin(x)+50", sin)
+	p.Legend.ThumbnailWidth = 0.5 * vg.Inch
+
+	p.X.Min = 0
+	p.X.Max = 10
+	p.Y.Min = 0
+	p.Y.Max = 100
+
+	err = p.Save(200, 200, "testdata/functions.png")
+	if err != nil {
+		log.Panic(err)
+	}
+}
