+++ plotter/functions_test.go
@@ -5,57 +5,10 @@
 package plotter_test
 
 import (
-	"image/color"
-	"log"
-	"math"
 	"testing"
 
-	"gonum.org/v1/plot"
 	"gonum.org/v1/plot/cmpimg"
-	"gonum.org/v1/plot/plotter"
-	"gonum.org/v1/plot/vg"
 )
-
-// ExampleFunction draws some functions.
-func ExampleFunction() {
-	quad := plotter.NewFunction(func(x float64) float64 { return x * x })
-	quad.Color = color.RGBA{B: 255, A: 255}
-
-	exp := plotter.NewFunction(func(x float64) float64 { return math.Pow(2, x) })
-	exp.Dashes = []vg.Length{vg.Points(2), vg.Points(2)}
-	exp.Width = vg.Points(2)
-	exp.Color = color.RGBA{G: 255, A: 255}
-
-	sin := plotter.NewFunction(func(x float64) float64 { return 10*math.Sin(x) + 50 })
-	sin.Dashes = []vg.Length{vg.Points(4), vg.Points(5)}
-	sin.Width = vg.Points(4)
-	sin.Color = color.RGBA{R: 255, A: 255}
-
-	p, err := plot.New()
-	if err != nil {
-		log.Panic(err)
-	}
-
-	p.Title.Text = "Functions"
-	p.X.Label.Text = "X"
-	p.Y.Label.Text = "Y"
-
-	p.Add(quad, exp, sin)
-	p.Legend.Add("x^2", quad)
-	p.Legend.Add("2^x", exp)
-	p.Legend.Add("10*sin(x)+50", sin)
-	p.Legend.ThumbnailWidth = 0.5 * vg.Inch
-
-	p.X.Min = 0
-	p.X.Max = 10
-	p.Y.Min = 0
-	p.Y.Max = 100
-
-	err = p.Save(200, 200, "testdata/functions.png")
-	if err != nil {
-		log.Panic(err)
-	}
-}
 
 func TestFunction(t *testing.T) {
 	cmpimg.CheckPlot(ExampleFunction, t, "functions.png")
