+++ plotter/scatter_test.go
@@ -5,86 +5,10 @@
 package plotter_test
 
 import (
-	"image/color"
-	"log"
 	"testing"
 
-	"golang.org/x/exp/rand"
-
-	"gonum.org/v1/plot"
 	"gonum.org/v1/plot/cmpimg"
-	"gonum.org/v1/plot/plotter"
-	"gonum.org/v1/plot/vg"
-	"gonum.org/v1/plot/vg/draw"
 )
-
-// ExampleScatter draws some scatter points, a line,
-// and a line with points.
-func ExampleScatter() {
-	rnd := rand.New(rand.NewSource(1))
-
-	// randomPoints returns some random x, y points
-	// with some interesting kind of trend.
-	randomPoints := func(n int) plotter.XYs {
-		pts := make(plotter.XYs, n)
-		for i := range pts {
-			if i == 0 {
-				pts[i].X = rnd.Float64()
-			} else {
-				pts[i].X = pts[i-1].X + rnd.Float64()
-			}
-			pts[i].Y = pts[i].X + 10*rnd.Float64()
-		}
-		return pts
-	}
-
-	n := 15
-	scatterData := randomPoints(n)
-	lineData := randomPoints(n)
-	linePointsData := randomPoints(n)
-
-	p, err := plot.New()
-	if err != nil {
-		log.Panic(err)
-	}
-	p.Title.Text = "Points Example"
-	p.X.Label.Text = "X"
-	p.Y.Label.Text = "Y"
-	p.Add(plotter.NewGrid())
-
-	s, err := plotter.NewScatter(scatterData)
-	if err != nil {
-		log.Panic(err)
-	}
-	s.GlyphStyle.Color = color.RGBA{R: 255, B: 128, A: 255}
-	s.GlyphStyle.Radius = vg.Points(3)
-
-	l, err := plotter.NewLine(lineData)
-	if err != nil {
-		log.Panic(err)
-	}
-	l.LineStyle.Width = vg.Points(1)
-	l.LineStyle.Dashes = []vg.Length{vg.Points(5), vg.Points(5)}
-	l.LineStyle.Color = color.RGBA{B: 255, A: 255}
-
-	lpLine, lpPoints, err := plotter.NewLinePoints(linePointsData)
-	if err != nil {
-		log.Panic(err)
-	}
-	lpLine.Color = color.RGBA{G: 255, A: 255}
-	lpPoints.Shape = draw.CircleGlyph{}
-	lpPoints.Color = color.RGBA{R: 255, A: 255}
-
-	p.Add(s, l, lpLine, lpPoints)
-	p.Legend.Add("scatter", s)
-	p.Legend.Add("line", l)
-	p.Legend.Add("line points", lpLine, lpPoints)
-
-	err = p.Save(200, 200, "testdata/scatter.png")
-	if err != nil {
-		log.Panic(err)
-	}
-}
 
 func TestScatter(t *testing.T) {
 	cmpimg.CheckPlot(ExampleScatter, t, "scatter.png")
