+++ plotutil/example_errorpoints_test.go
