+++ vg/vgsvg/example_vgsvg_test.go
@@ -1,0 +1,33 @@
+// Copyright ©2019 The Gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package vgsvg_test
+
+import (
+	"gonum.org/v1/plot"
+	"gonum.org/v1/plot/plotter"
+	"gonum.org/v1/plot/vg"
+	"log"
+)
+
+func Example() {
+	p, err := plot.New()
+	if err != nil {
+		log.Fatalf("could not create plot: %v", err)
+	}
+	p.Title.Text = "Scatter plot"
+	p.X.Label.Text = "X"
+	p.Y.Label.Text = "Y"
+
+	scatter, err := plotter.NewScatter(plotter.XYs{{1, 1}, {0, 1}, {0, 0}})
+	if err != nil {
+		log.Fatalf("could not create scatter: %v", err)
+	}
+	p.Add(scatter)
+
+	err = p.Save(5*vg.Centimeter, 5*vg.Centimeter, "testdata/scatter.svg")
+	if err != nil {
+		log.Fatalf("could not save SVG plot: %v", err)
+	}
+}
