+++ vg/vgsvg/vgsvg_test.go
@@ -7,7 +7,6 @@ package vgsvg_test
 import (
 	"bytes"
 	"io/ioutil"
-	"log"
 	"testing"
 
 	"gonum.org/v1/plot"
@@ -17,27 +16,6 @@ import (
 	"gonum.org/v1/plot/vg/draw"
 	"gonum.org/v1/plot/vg/vgsvg"
 )
-
-func Example() {
-	p, err := plot.New()
-	if err != nil {
-		log.Fatalf("could not create plot: %v", err)
-	}
-	p.Title.Text = "Scatter plot"
-	p.X.Label.Text = "X"
-	p.Y.Label.Text = "Y"
-
-	scatter, err := plotter.NewScatter(plotter.XYs{{1, 1}, {0, 1}, {0, 0}})
-	if err != nil {
-		log.Fatalf("could not create scatter: %v", err)
-	}
-	p.Add(scatter)
-
-	err = p.Save(5*vg.Centimeter, 5*vg.Centimeter, "testdata/scatter.svg")
-	if err != nil {
-		log.Fatalf("could not save SVG plot: %v", err)
-	}
-}
 
 func TestSVG(t *testing.T) {
 	cmpimg.CheckPlot(Example, t, "scatter.svg")
